Refactored dpos constants and updated imports
