+++ packages/views/screens/managers/monitor/delegates/overview/delegatesOverview.js
@@ -1,7 +1,7 @@
 // istanbul ignore file
 import React from 'react';
 // import { fromRawLsk } from '@token/utilities/lsk';
-import { ROUND_LENGTH } from '@dpos/configuration/delegates';
+import { ROUND_LENGTH } from '@dpos/validator/constants';
 import { useTheme } from '@views/utilities/theme';
 import { getColorPalette } from '@views/basics/charts/chartOptions';
 import Box from '@basics/box';
