+++ src/modules/dpos/validator/constants/index.js
@@ -1,0 +1,1 @@
+export * from './delegates';
